Expose ExceptionTracker as public API - closes #36
